in progress commit, fix shaders

+++ banan_buffer.h
@@ -24,10 +24,10 @@ namespace Banan {
         VkDescriptorBufferInfo descriptorInfo(VkDeviceSize size = VK_WHOLE_SIZE, VkDeviceSize offset = 0);
         VkResult invalidate(VkDeviceSize size = VK_WHOLE_SIZE, VkDeviceSize offset = 0);
 
-        void writeToIndex(void* data, int index);
-        VkResult flushIndex(int index);
-        VkDescriptorBufferInfo descriptorInfoForIndex(int index);
-        VkResult invalidateIndex(int index);
+        void writeToIndex(void* data, size_t index);
+        VkResult flushIndex(size_t index);
+        VkDescriptorBufferInfo descriptorInfoForIndex(size_t index);
+        VkResult invalidateIndex(size_t index);
 
         VkBuffer getBuffer();
         void* getMappedMemory();
